yana-comu-second/src/requirements.txt
accelerate==1.8.1
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
dotenv==0.9.9
fastapi==0.115.14
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
SQLAlchemy==2.0.28
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.4
line-bot-sdk==3.17.1
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numpy==2.2.6
ollama==0.5.1
packaging==25.0
pillow==11.0.0
propcache==0.3.2
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==80.9.0
simpleaudio==1.0.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.0
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
